requirements.txt
git+https://github.com/bbelderbos/Carbon-API.git
fastapi
uvicorn
aiofiles
bcrypt
passlib
python-dotenv
python-jose
python-multipart
requests
tortoise-orm
